+++ client/quantum_serverless/core/provider.py
@@ -53,7 +53,7 @@ from quantum_serverless.core.job import (
     GatewayJobClient,
     BaseJobClient,
 )
-from quantum_serverless.core.program import QiskitPattern
+from quantum_serverless.core.pattern import QiskitPattern
 from quantum_serverless.core.tracing import _trace_env_vars
 from quantum_serverless.exception import QuantumServerlessException
 from quantum_serverless.utils import JsonSerializable
