+++ client/tests/core/test_pattern.py
@@ -7,7 +7,7 @@ from testcontainers.compose import DockerCompose
 from quantum_serverless import QuantumServerless, BaseProvider
 from quantum_serverless.core import ComputeResource
 from quantum_serverless.core.job import Job
-from quantum_serverless.core.program import QiskitPattern
+from quantum_serverless.core.pattern import QiskitPattern
 from tests.utils import wait_for_job_client, wait_for_job_completion
 
 resources_path = os.path.join(
